refactor splitting between deployment and testing
