+++ utils/DeploymentUtil.js
@@ -1,0 +1,653 @@
+// Whole-script strict mode syntax
+"use strict";
+
+/**
+MIT License
+
+Copyright (c) 2020 Openlaw
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+ */
+
+const {
+  SHARES,
+  LOOT,
+  ETH_TOKEN,
+  sha3,
+  toBN,
+} = require('./DaoFactory');
+
+const deployDefaultDao = async (owner) => {
+  return await deployDao(null, { owner });
+};
+
+const deployDefaultNFTDao = async (owner) => {
+  const { dao, adapters, extensions, testContracts } = await deployDao(null, {
+    owner,
+    deployTestTokens: true,
+    finalize: false,
+  });
+
+  await dao.finalizeDao({ from: owner });
+  return {
+    dao: dao,
+    adapters: adapters,
+    extensions: extensions,
+    testContracts: testContracts,
+  };
+};
+
+const deployDaoWithOffchainVoting = async (owner, newMember) => {
+  const {
+    dao,
+    adapters,
+    extensions,
+    testContracts,
+    votingHelpers,
+  } = await deployDao(null, {
+    owner,
+    offchainVoting: true,
+    deployTestTokens: true,
+    finalize: false,
+  });
+
+  await dao.potentialNewMember(newMember, {
+    from: owner,
+  });
+  await extensions.bank.addToBalance(newMember, SHARES, 1, {
+    from: owner,
+  });
+
+  await dao.finalizeDao({ from: owner });
+
+  adapters["voting"] = votingHelpers.offchainVoting;
+
+  return {
+    dao: dao,
+    adapters: adapters,
+    extensions: extensions,
+    testContracts: testContracts,
+    votingHelpers: votingHelpers,
+  };
+};
+
+const deployDaoWithBatchVoting = async (owner, newMember) => {
+  const { dao, adapters, extensions, votingHelpers } = await deployDao(null, {
+    owner,
+    deployTestTokens: false,
+    batchVoting: true,
+    finalize: false,
+  });
+
+  await dao.potentialNewMember(newMember, {
+    from: owner,
+  });
+
+  await extensions.bank.addToBalance(newMember, SHARES, 1, {
+    from: owner,
+  });
+
+  await dao.finalizeDao({ from: owner });
+
+  adapters["voting"] = adapters.batchVoting;
+
+  return {
+    dao: dao,
+    adapters: adapters,
+    extensions: extensions,
+    votingHelpers: votingHelpers,
+  };
+};
+
+const deployDao = async (options) => {
+  const {deployFunction, owner, deployTestTokens, finalize, DaoRegistry, BankExtension, BankFactory, NFTExtension, NFTCollectionFactory, TestToken1, TestToken2, Multicall, PixelNFT, OLToken} = options;
+
+  let identityDao = await deployFunction(DaoRegistry);
+  let identityBank = await deployFunction(BankExtension);
+  let bankFactory = await deployFunction(BankFactory, [identityBank.address]);
+  let identityNft = await deployFunction(NFTExtension);
+  let nftFactory = await deployFunction(NFTCollectionFactory, [identityNft.address]);
+  
+  const { dao, daoFactory } = await cloneDao({...options, identityDao, name: 'test-dao'});
+
+  await bankFactory.createBank(options.maxExternalTokens);
+  let pastEvent;
+  while (pastEvent === undefined) {
+    let pastEvents = await bankFactory.getPastEvents();
+    pastEvent = pastEvents[0];
+  }
+
+  const { bankAddress } = pastEvent.returnValues;
+  const bankExtension = await BankExtension.at(bankAddress);
+  const creator = await dao.getMemberAddress(1);
+  await dao.addExtension(sha3("bank"), bankExtension.address, creator);
+
+  await nftFactory.createNFTCollection();
+  pastEvent = undefined;
+  while (pastEvent === undefined) {
+    let pastEvents = await nftFactory.getPastEvents();
+    pastEvent = pastEvents[0];
+  }
+
+  const { nftCollAddress } = pastEvent.returnValues;
+  const nftExtension = await NFTExtension.at(nftCollAddress);
+  await dao.addExtension(sha3("nft"), nftCollAddress, creator);
+
+  const extensions = { bank: bankExtension, nft: nftExtension };
+
+  const { adapters } = await addDefaultAdapters({
+    dao,
+    options,
+    daoFactory
+  });
+
+  const votingAddress = await dao.getAdapterAddress(sha3("voting"));
+  const votingHelpers = {
+    snapshotProposalContract: null,
+    handleBadReporterAdapter: null,
+    offchainVoting: null,
+    batchVoting: null,
+  };
+
+  if (options.isOffchainVoting) {
+    const {
+      offchainVoting,
+      handleBadReporterAdapter,
+      snapshotProposalContract,
+    } = await configureOffchainVoting(
+      {
+        ...options, 
+        dao,
+        daoFactory,
+        votingAddress,
+        bankAddress
+      });
+    votingHelpers.offchainVoting = offchainVoting;
+    votingHelpers.handleBadReporterAdapter = handleBadReporterAdapter;
+    votingHelpers.snapshotProposalContract = snapshotProposalContract;
+  } else if (options.isBatchVoting) {
+    const {
+      batchVoting,
+      snapshotProposalContract,
+    } = await configureBatchVoting({
+      ...options,
+      dao,
+      daoFactory
+    });
+    
+    votingHelpers.batchVoting = batchVoting;
+    votingHelpers.snapshotProposalContract = snapshotProposalContract;
+  }
+
+  // deploy test token contracts (for testing convenience)
+  const testContracts = {
+    oltToken: null,
+    testToken1: null,
+    testToken2: null,
+    multicall: null,
+    pixelNFT: null,
+  };
+
+  if (deployTestTokens) {
+    testContracts.testToken1 = await deployFunction(TestToken1, [1000000]);
+    testContracts.testToken2 = await deployFunction(TestToken2, [1000000]);
+    testContracts.multicall = await deployFunction(Multicall);
+    testContracts.pixelNFT = await deployFunction(PixelNFT, [100]);
+    testContracts.oltToken = await deployFunction(OLToken,[toBN("1000000000000000000000000")]);
+  }
+
+  if (finalize) {
+    await dao.finalizeDao({ from: owner });
+  }
+
+  return {
+    dao: dao,
+    adapters: adapters,
+    extensions: extensions,
+    testContracts: testContracts,
+    votingHelpers: votingHelpers,
+  };
+};
+
+const prepareAdapters = async ({deployFunction, VotingContract, ConfigurationContract, TributeContract, TributeNFTContract,  DistributeContract, RagequitContract, ManagingContract, FinancingContract, OnboardingContract, GuildKickContract, DaoRegistryAdapterContract, BankAdapterContract, NFTAdapterContract, CouponOnboardingContract}) => {
+  
+  let voting, configuration, ragequit, managing, financing, onboarding, guildkick, daoRegistryAdapter, bankAdapter, nftAdapter, couponOnboarding, tribute, distribute, tributeNFT;
+
+  voting = await deployFunction(VotingContract);
+  configuration = await deployFunction(ConfigurationContract);
+  ragequit = await deployFunction(RagequitContract);
+  managing = await deployFunction(ManagingContract);
+  financing = await deployFunction(FinancingContract);
+  onboarding = await deployFunction(OnboardingContract);
+  guildkick = await deployFunction(GuildKickContract);
+  daoRegistryAdapter = await deployFunction(DaoRegistryAdapterContract);
+  bankAdapter = await deployFunction(BankAdapterContract);
+  nftAdapter = await deployFunction(NFTAdapterContract);
+  couponOnboarding = await deployFunction(CouponOnboardingContract, [1]);
+  tribute = await deployFunction(TributeContract);
+  distribute = await deployFunction(DistributeContract);
+  tributeNFT = await deployFunction(TributeNFTContract);
+
+  return {
+    voting,
+    configuration,
+    ragequit,
+    guildkick,
+    managing,
+    financing,
+    onboarding,
+    daoRegistryAdapter,
+    bankAdapter,
+    nftAdapter,
+    couponOnboarding,
+    tribute,
+    distribute,
+    tributeNFT,
+  };
+};
+
+const createIdentityDao = async (options) => {
+  let {DaoRegistry} = options;
+
+  return await DaoRegistry.new({
+    from: options.owner,
+    gasPrice: toBN("0"),
+  });
+};
+
+const addDefaultAdapters = async ({
+  dao,
+  options,
+  daoFactory,
+  nftAddr
+}) => {
+  const {
+    voting,
+    configuration,
+    ragequit,
+    guildkick,
+    managing,
+    financing,
+    onboarding,
+    daoRegistryAdapter,
+    bankAdapter,
+    nftAdapter,
+    couponOnboarding,
+    tribute,
+    distribute,
+    tributeNFT,
+  } = await prepareAdapters(options);
+
+  let {BankExtension, NFTExtension} = options;
+
+  const bankAddress = await dao.getExtensionAddress(sha3("bank"));
+  const bankExtension = await BankExtension.at(bankAddress);
+
+  const nftExtAddr = await dao.getExtensionAddress(sha3("nft"));
+  const nftExtension = await NFTExtension.at(nftExtAddr);
+
+  await configureDao({
+    daoFactory,
+    dao,
+    ragequit,
+    guildkick,
+    managing,
+    financing,
+    onboarding,
+    daoRegistryAdapter,
+    bankAdapter,
+    nftAdapter,
+    voting,
+    configuration,
+    couponOnboarding,
+    tribute,
+    distribute,
+    tributeNFT,
+    nftAddr,
+    bankExtension,
+    nftExtension,
+    ... options
+  });
+
+  return {
+    dao,
+    adapters: {
+      voting,
+      configuration,
+      ragequit,
+      guildkick,
+      managing,
+      financing,
+      onboarding,
+      daoRegistryAdapter,
+      bankAdapter,
+      nftAdapter,
+      couponOnboarding,
+      tribute,
+      distribute,
+      tributeNFT,
+    },
+  };
+};
+
+const configureDao = async ({
+  daoFactory,
+  dao,
+  ragequit,
+  guildkick,
+  managing,
+  financing,
+  onboarding,
+  daoRegistryAdapter,
+  bankAdapter,
+  bankExtension,
+  nftExtension,
+  nftAdapter,
+  voting,
+  configuration,
+  couponOnboarding,
+  tribute,
+  distribute,
+  tributeNFT,
+  unitPrice,
+  maxChunks,
+  nbShares,
+  tokenAddr,
+  votingPeriod,
+  gracePeriod,
+  couponCreatorAddress
+}) => {
+  await daoFactory.addAdapters(dao.address, [
+    entryDao("voting", voting, {}),
+    entryDao("configuration", configuration, {
+      SUBMIT_PROPOSAL: true,
+      SET_CONFIGURATION: true,
+    }),
+    entryDao("ragequit", ragequit, {}),
+    entryDao("guildkick", guildkick, {
+      SUBMIT_PROPOSAL: true,
+    }),
+    entryDao("managing", managing, {
+      SUBMIT_PROPOSAL: true,
+      REPLACE_ADAPTER: true,
+    }),
+    entryDao("financing", financing, {
+      SUBMIT_PROPOSAL: true,
+    }),
+    entryDao("onboarding", onboarding, {
+      SUBMIT_PROPOSAL: true,
+      UPDATE_DELEGATE_KEY: true,
+      NEW_MEMBER: true,
+    }),
+    entryDao("coupon-onboarding", couponOnboarding, {
+      SUBMIT_PROPOSAL: false,
+      ADD_TO_BALANCE: true,
+      UPDATE_DELEGATE_KEY: false,
+      NEW_MEMBER: true,
+    }),
+    entryDao("daoRegistry", daoRegistryAdapter, {
+      UPDATE_DELEGATE_KEY: true,
+    }),
+    entryDao("nft", nftAdapter, {}),
+    entryDao("bank", bankAdapter, {}),
+    entryDao("tribute", tribute, {
+      SUBMIT_PROPOSAL: true,
+      NEW_MEMBER: true,
+    }),
+    entryDao("tribute-nft", tributeNFT, {
+      SUBMIT_PROPOSAL: true,
+      NEW_MEMBER: true,
+    }),
+    entryDao("distribute", distribute, {
+      SUBMIT_PROPOSAL: true,
+    }),
+  ]);
+
+  await daoFactory.configureExtension(dao.address, bankExtension.address, [
+    entryBank(ragequit, {
+      INTERNAL_TRANSFER: true,
+      SUB_FROM_BALANCE: true,
+      ADD_TO_BALANCE: true,
+    }),
+    entryBank(guildkick, {
+      INTERNAL_TRANSFER: true,
+      SUB_FROM_BALANCE: true,
+      ADD_TO_BALANCE: true,
+    }),
+    entryBank(bankAdapter, {
+      WITHDRAW: true,
+      SUB_FROM_BALANCE: true,
+      UPDATE_TOKEN: true,
+    }),
+    entryBank(onboarding, {
+      ADD_TO_BALANCE: true,
+      SUB_FROM_BALANCE: true,
+    }),
+    entryBank(couponOnboarding, {
+      ADD_TO_BALANCE: true,
+    }),
+    entryBank(financing, {
+      ADD_TO_BALANCE: true,
+      SUB_FROM_BALANCE: true,
+    }),
+    entryBank(tribute, {
+      ADD_TO_BALANCE: true,
+      SUB_FROM_BALANCE: true,
+      REGISTER_NEW_TOKEN: true,
+    }),
+    entryBank(distribute, {
+      INTERNAL_TRANSFER: true,
+    }),
+    entryBank(tributeNFT, {
+      ADD_TO_BALANCE: true,
+    }),
+  ]);
+
+  await daoFactory.configureExtension(dao.address, nftExtension.address, [
+    entryNft(tributeNFT, {
+      COLLECT_NFT: true,
+    }),
+    entryNft(nftAdapter, {
+      COLLECT_NFT: true,
+    }),
+  ]);
+  
+  await onboarding.configureDao(
+    dao.address,
+    SHARES,
+    unitPrice,
+    nbShares,
+    maxChunks,
+    tokenAddr
+  );
+
+  await onboarding.configureDao(
+    dao.address,
+    LOOT,
+    unitPrice,
+    nbShares,
+    maxChunks,
+    tokenAddr
+  );
+  await couponOnboarding.configureDao(
+    dao.address,
+    couponCreatorAddress,
+    SHARES
+  );
+
+  await voting.configureDao(dao.address, votingPeriod, gracePeriod);
+  await tribute.configureDao(dao.address, SHARES);
+  await tribute.configureDao(dao.address, LOOT);
+  await tributeNFT.configureDao(dao.address);
+};
+
+const cloneDao = async ({identityDao, owner, deployFunction, DaoRegistry, DaoFactory, name}) => {
+  let daoFactory = await deployFunction(DaoFactory, [identityDao.address], owner);
+
+  if(owner) {
+    await daoFactory.createDao(name, ETH_TOKEN, {from: owner});
+  } else {
+    await daoFactory.createDao(name, ETH_TOKEN);
+  }
+  
+  // checking the gas usaged to clone a contract
+  let pastEvents = await daoFactory.getPastEvents();
+  let { _address, _name } = pastEvents[0].returnValues;
+  let newDao = await DaoRegistry.at(_address);
+  return { dao: newDao, daoFactory, daoName: _name };
+};
+
+const entryNft = (contract, flags) => {
+  const values = [
+    flags.WITHDRAW_NFT,
+    flags.COLLECT_NFT,
+    flags.INTERNAL_TRANSFER,
+  ];
+
+  const acl = entry(values);
+
+  return {
+    id: sha3("n/a"),
+    addr: contract.address,
+    flags: acl,
+  };
+};
+
+const entryBank = (contract, flags) => {
+  const values = [
+    flags.ADD_TO_BALANCE,
+    flags.SUB_FROM_BALANCE,
+    flags.INTERNAL_TRANSFER,
+    flags.WITHDRAW,
+    flags.EXECUTE,
+    flags.REGISTER_NEW_TOKEN,
+    flags.REGISTER_NEW_INTERNAL_TOKEN,
+  ];
+
+  const acl = entry(values);
+
+  return {
+    id: sha3("n/a"),
+    addr: contract.address,
+    flags: acl,
+  };
+};
+
+const entryDao = (name, contract, flags) => {
+  const values = [
+    flags.REPLACE_ADAPTER,
+    flags.SUBMIT_PROPOSAL,
+    flags.UPDATE_DELEGATE_KEY,
+    flags.SET_CONFIGURATION,
+    flags.ADD_EXTENSION,
+    flags.REMOVE_EXTENSION,
+    flags.NEW_MEMBER,
+  ];
+
+  const acl = entry(values);
+
+  return {
+    id: sha3(name),
+    addr: contract.address,
+    flags: acl,
+  };
+};
+
+const entry = (values) => {
+  return values
+    .map((v, idx) => (v === true ? 2 ** idx : 0))
+    .reduce((a, b) => a + b);
+};
+
+const configureOffchainVoting = async ({
+  dao,
+  daoFactory,
+  chainId,
+  votingAddress,
+  bankAddress,
+  votingPeriod,
+  gracePeriod,
+  deployFunction,
+  SnapshotProposalContract,
+  KickBadReporterAdapter,
+  OffchainVotingContract
+}) => {
+  let snapshotProposalContract = await deployFunction(SnapshotProposalContract, [chainId]);
+  let handleBadReporterAdapter = await deployFunction(KickBadReporterAdapter);
+  let offchainVoting = await deployFunction(OffchainVotingContract, [votingAddress, snapshotProposalContract.address, handleBadReporterAdapter.address]);
+
+  await daoFactory.updateAdapter(
+    dao.address,
+    entryDao("voting", offchainVoting, {})
+  );
+
+  await dao.setAclToExtensionForAdapter(
+    bankAddress,
+    offchainVoting.address,
+    entryBank(offchainVoting, {
+      ADD_TO_BALANCE: true,
+      SUB_FROM_BALANCE: true,
+      INTERNAL_TRANSFER: true,
+    }).flags
+  );
+  await offchainVoting.configureDao(dao.address, votingPeriod, gracePeriod, 10);
+
+  return { offchainVoting, snapshotProposalContract, handleBadReporterAdapter };
+};
+
+const configureBatchVoting = async ({
+  dao,
+  daoFactory,
+  deployFunction,
+  chainId,
+  SnapshotProposalContract,
+  BatchVotingContract
+}
+) => {
+  let snapshotProposalContract, batchVoting;
+  snapshotProposalContract = await deployFunction(SnapshotProposalContract, [chainId]);
+  batchVoting = await options.deployFunction(BatchVotingContract, [snapshotProposalContract.address]);
+  
+  await daoFactory.updateAdapter(
+    dao.address,
+    entryDao("voting", batchVoting, {}),
+    { from: options.owner }
+  );
+
+  await batchVoting.configureDao(dao.address, options.votingPeriod, options.gracePeriod, {
+    from: options.owner,
+  });
+
+  return { batchVoting, snapshotProposalContract };
+};
+
+module.exports = {
+  deployDao,
+  deployDefaultDao,
+  deployDefaultNFTDao,
+  deployDaoWithBatchVoting,
+  deployDaoWithOffchainVoting,
+  createIdentityDao,
+  cloneDao,
+  prepareAdapters,
+  addDefaultAdapters,
+  entry,
+  entryBank,
+  entryDao
+};
